Flytt valideringstester til korrekt fil
